Editor

Description

The code for mapping user actions via keyboard to operations on the zipper and to the display.

Documentation

data Mode

Constructors

Insert

Edit

data TedState

Constructors

TedState

Fields

mode :: Mode

offset :: Int

lines above top of the window

text :: Zipper

path :: String

file path to save to

clip :: Text

clipboard contents

type MonadEditor a = StateT TedState Curses a

startState :: Zipper -> String -> TedState

runEditor :: TedState -> Curses ()

eventLoop :: MonadEditor ()

handleEvent :: Window -> Event -> MonadEditor ()

copy :: MonadEditor ()

paste :: Window -> MonadEditor ()

editEvent :: Window -> Event -> MonadEditor ()

Handles an edit mode event by making the corresponding zipper update and displaying it

insertEvent :: Window -> Event -> MonadEditor ()

Handles an insert mode event by making the corresponding zipper update and displaying it

act :: Int -> Key -> Zipper -> Zipper

Maps a key press to the corresponding operation on the zipper

moveCursorInt :: Int -> Int -> Update ()

screenSizeInt :: Curses (Int, Int)

getCursorInt :: Window -> Curses (Int, Int)

draw :: Int -> Zipper -> Curses ()

Draws a zipper to the terminal. Skips offset o lines from the beginning of the zipper

tedRender :: Window -> (Zipper -> Zipper) -> MonadEditor ()

Updates the display based on the state of the zipper.

arrowToMove :: Int -> Zipper -> Key -> (Int, Int)

Converts an arrow key press to the (y,x) go-relative coordinates.

arrowKeys :: [Key]

isArrowKey :: Key -> Bool

toggleMode :: Mode -> Mode

write :: FilePath -> Zipper -> IO ()