Lib

Description

Documentation

appN :: Integral n => (a -> a) -> n -> a -> a

Successively apply a function n times.

appFix :: (Integral n, Eq a) => (a -> a) -> n -> a -> a

Successively apply a function n times or until it reaches a fixpoint

both :: (a -> b) -> (a, a) -> (b, b)

Apply a function to both elements of a tuple

concatWith :: (Monoid a, Foldable t) => a -> t a -> a

Concatenate the elements of a container with an interspersed element

concatS :: (Monoid a, Foldable t) => t a -> a

Concatenate the elements of a container. For use with Seq since I couldn't use the prelude version on it

seqUnsnoc :: Seq a -> (Seq a, a)

Split a Seq a into (init, last)

seqUncons :: Seq a -> (a, Seq a)

Split a Seq a into (head, tail)

pairDiff :: (Num a, Num b) => (a, b) -> (a, b) -> (a, b)

(+++) :: Text -> Text -> Text

Append operator for Text

roundUp :: Integral a => a -> a -> a

Rounds up to a positive multiple of some base