Zipper

Description

The functions and data types for the creating and manipulating the core data structure of the editor.

Documentation

type Position = (Int, Int)

data Zipper

A 2-d zipper representing the contents of a text file with newlines stripped. A paragraph is represented by the Text to the left of the currently-selected contents, and the Text to its right.

Constructors

Zipper

Fields

above :: Seq Text

The paragraphs of text above the current line

left :: Text

The text left and above the current selection

selection :: Text

The current selection

right :: Text

The text right and below the current selection

below :: Seq Text

The paragraphs of text below the current line

emptyZipper :: Zipper

makeZipper :: Seq Text -> Text -> Text -> Text -> Seq Text -> Zipper

A not-so-smart constructor for a Zipper. If the underlying representation of a Zipper is extended, old examples and test cases might not have to be changed.

paragraph :: Text -> Text -> Text -> Zipper

Makes a Zipper with no lines above or below.

fromLine :: Int -> Text -> Zipper

Makes the (left, selection, right) portion of a Zipper at the specified column from a line of text.

>>> fromLine 0 "123456789" == paragraph "" "1" "23456789"
True

splitThree :: Int -> Seq Text -> (Seq Text, Text, Seq Text)

Split a Seq Text into (above, selection, below) at the specified row (0-indexed).

fromText :: Position -> Text -> Zipper

Makes a Zipper from Text by splitting on newlines.

>>> fromText (0,0) "1\n2\n3\n" == makeZipper [] "" "1" "" ["2","3"]
True

combineLine :: Zipper -> Text

toLines :: Zipper -> Seq Text

toText :: Zipper -> Text

Converts a Zipper into Text by interspersing newlines between all lines. Obivated in editor by toText3 and only used for testing.

>>> let text = "1\n2\n3\n4\n" in let z = fromText (1,1) text in text == toText z
True

parLen :: Zipper -> Int

parLines :: Int -> Zipper -> Int

toText3 :: Int -> Int -> Int -> Zipper -> (Text, Text, Text)

Converts the visible portion of a Zipper into the text before the selection, the selection, and the text after the selection

splitStart :: Int -> Zipper -> (Text, Text)

takeTop :: Int -> Int -> Zipper -> Seq Text

Takes n lines of length w (or ending in a newline) from the text which precedes the first line of the current selection

splitEnd :: Int -> Zipper -> (Text, Text)

takeBottom :: Int -> Int -> Zipper -> Seq Text

Takes n lines of length w (or ending in a newline) from the text which follows the last line of the current selection

atBottom :: Zipper -> Bool

atTop :: Zipper -> Bool

atEnd :: Zipper -> Bool

upPar :: Zipper -> Zipper

Moves the cursor up one row. Does nothing if the cursor is already at the topmost position.

>>> upPar (makeZipper ["0", "1"] "1" "2" "3" ["b"]) == makeZipper ["0"] "" "1" "" ["123", "b"]
True

>>> upPar (makeZipper ["hi"] "1" "2" "3" ["b"]) == makeZipper [] "" "h" "i" ["123", "b"]
True

>>> upPar (makeZipper ["hello"] "1" "2" "3" []) == makeZipper [] "" "h" "ello" ["123"]
True

upPars :: Int -> Zipper -> Zipper

Moves up n rows

downPar :: Zipper -> Zipper

Moves the cursor down one row. Does nothing if the cursor is already at the bottommost position.

>>> downPar (makeZipper ["0"] "" "1" "2" ["b", "c"]) == makeZipper ["0", "12"] "" "b" "" ["c"]
True

>>> downPar (makeZipper [] "" "!" "" ["single"]) == makeZipper ["!"] "" "s" "ingle" []
True

downPars :: Int -> Zipper -> Zipper

Moves down n rows

stepLeft :: Zipper -> Zipper

Moves the cursor left one column or to the end of the previous paragraph if at the leftmost position

>>> stepLeft numbers == paragraph "123" "4" "56789"
True

>>> goLeft 20 numbers == paragraph "" "1" "23456789"
True

>>> toText (stepLeft numbers) == toText numbers
True

goLeft :: Int -> Zipper -> Zipper

Moves left n times

stepRight :: Zipper -> Zipper

Moves the cursor right one column. Does nothing if the cursor is already at the rightmost position.

>>> stepRight numbers == paragraph "12345" "6" "789"
True

>>> goRight 20 numbers == paragraph "123456789" "" ""
True

>>> toText (stepRight numbers) == toText numbers
True

goRight :: Int -> Zipper -> Zipper

Moves right n columns

go :: Int -> (Int, Int) -> Zipper -> Zipper

Performs a move on a zipper (row, column). Vertically relative, horizontally relative to the x position after the vertical move.

stepDown :: Int -> Zipper -> Zipper

Moves down one visual line (e.g. n characters, the length of a line on the screen)

>>> stepDown 5 (paragraph "" "l" "ine line") == paragraph "line " "l" "ine"
True

>>> stepDown 20 (Zipper [] "between para" "" "" [""]) == Zipper ["between para"] "" "" "" []
True

>>> stepDown 20 (Zipper [] "between para" "" "" ["!"]) == Zipper ["between para"] "!" "" "" []
True

>>> stepDown 20 (Zipper [] "be" "t" "ween" ["!!!"]) == Zipper ["between"] "!!" "!" "" []
True

>>> stepDown 4 (Zipper [] "lin" "e" "l" ["not here"]) == Zipper [] "linel" "" "" ["not here"]
True

goDown :: Int -> Int -> Zipper -> Zipper

stepUp :: Int -> Zipper -> Zipper

Moves up one visual line (e.g. n characters, the length of a line on the screen)

>>> stepUp 2 (Zipper [] "abcde" "f" "" []) == Zipper [] "abc" "d" "ef" []
True

>>> stepUp 2 (Zipper [""] "" "a" "b" []) == Zipper [] "" "" "" ["ab"]
True

>>> stepUp 10 (Zipper ["abcdef"] "" "g" "h" []) == Zipper [] "" "a" "bcdef" ["gh"]
True

>>> stepUp 10 (Zipper ["ab"] "cdef" "g" "h" []) == Zipper [] "ab" "" "" ["cdefgh"]
True

goUp :: Int -> Int -> Zipper -> Zipper

selectionPos :: Int -> Zipper -> (Position, Position)

cursorStart :: Int -> Zipper -> Position

Returns a 0-indexed (row, column) cursor position based on lines of length n

>>> cursorStart 10 emptyZipper == (0,0)
True

>>> cursorStart 10 $ makeZipper ["0", "1", "2", "3"] "01234" "5" "6789" ["below"]
(4,5)

>>> cursorStart 2 $ makeZipper ["0", "1", "2", "3"] "01234" "56" "6789" ["below"]
(6,1)

cursorEnd :: Int -> Zipper -> Position

Returns a 0-indexed (row, column) cursor position based on lines of length n

>>> cursorEnd 10 emptyZipper == (0,0)
True

>>> cursorEnd 10 $ makeZipper ["0", "1", "2", "3"] "01234" "5" "6789" ["below"]
(4,5)

>>> cursorEnd 2 $ makeZipper ["0", "1", "2", "3"] "01234" "56" "789" ["below"]
(7,0)

lsAbove :: Int -> Zipper -> Int

cursorCol :: Int -> Zipper -> Position

Returns a 0-indexed (colStart, colEnd) cursor position based on lines of length n

cursorRow :: Int -> Zipper -> Position

Returns a 0-indexed (rowStart, rowEnd) cursor position based on lines of length n

split :: Zipper -> Zipper

Creates a new line from the current selection

>>> split (paragraph "moves up" "d" "oesn't") == makeZipper ["moves up"] "" "d" "oesn't" []
True

>>> split (paragraph "" "stays" "put") == makeZipper [""] "" "stays" "put" []
True

backspace :: Zipper -> Zipper

Removes a character from left

>>> backspace (paragraph "abc" "def" "ghi") == paragraph "ab" "def" "ghi"
True

>>> backspace (Zipper ["moves"] "" "u" "p" []) == Zipper [] "moves" "u" "p" []
True

delete :: Zipper -> Zipper

Deletes the current selection

>>> delete emptyZipper == emptyZipper
True

delete (paragraph "" A "") == emptyZipper True

>>> delete (paragraph "abc" "d" "efg") == paragraph "abc" "e" "fg"
True

>>> delete (paragraph "abc" "de" "fg") == paragraph "abc" "f" "g"
True

>>> delete (Zipper [] "abc" "d" "" ["e"]) == Zipper [] "abc" "" "" ["e"]
True

>>> delete (Zipper [] "text" "" "" ["moves up"]) == Zipper [] "text" "m" "oves up" []
True

appendChar :: Char -> Zipper -> Zipper

Appends a character to the left of the selection

appendText :: Text -> Zipper -> Zipper

Appends text to the left of the selection

selectSentence :: Zipper -> Zipper

Extends the cursor to include the sentence which encloses the current selection. Does not handle acronyms well.

>>> selectSentence zSentences == zSentences'
True

selectParagraph :: Zipper -> Zipper

toSelection :: (Text -> Text) -> Zipper -> Zipper

Applies a function to the selection

toUpper :: Zipper -> Zipper

toLower :: Zipper -> Zipper

sentences :: String -> [(Text, Int)]

selectEnclosed :: Text -> Zipper -> Zipper

Selects the text which is enclosed by some string s

selectWord :: Zipper -> Zipper

posDiff :: Int -> Zipper -> Zipper -> (Int, Int)

numbers :: Zipper

zSentences :: Zipper

zSentences' :: Zipper

exampleZipper :: Zipper

A zipper which can be manipulated in ghci in case a reviewer has trouble building the project.

ez :: Zipper